[1] Use of raw partition as log device and data device is enabled for Linux     platform. Linux has a slight deviation in philosophy from other UNIX     with respect to the handling of character oriented (raw) interface to     block device. It uses the same interface in VFS. So, while in mach we     need to use /dev/rz0g and /dev/rrz0g to represent the block and     character interface to the g partition of the disk rz, in Linux, we use     /dev/sda3 for _both_ the block and char I/F for partition 3 of first     scsi (sda). [3] Codes requiring libplumber are comment out unless the macro HAS_PLUMBER     is specifically defined. Those codes are only error checking codes. The     macro HAS_PLUMBER should be defined in makefile for platform that has     libplumber. (detailed file-by-file explanation is also posted on coda.general on Mar     11, 97 and Mar 19, 97) -- Clement

+++ rvm-src/rds/rds_zap.c
@@ -60,7 +60,7 @@ rds_zap_heap(DevName, DevLength, startAddr, staticLength, heapLength, nlists, ch
 {
     rvm_region_def_t regions[2], *loadregions;
     rvm_tid_t *tid;
-    int n_loadregions;
+    unsigned long n_loadregions;
     rvm_return_t rvmret;
     
     regions[0].length = heapLength;
